sql: add EXPLAIN (VEC) output to the stmt bundle
This commit refactors the code of `EXPLAIN (VEC)` to be reused to add
`(VEC)` and `(VEC, VERBOSE)` flavours to the statement bundle. The code
to print the diagram is moved into `colflow` package, but other changes
are minor.

The addition of these diagrams is performed on the best effort basis
(meaning that if an error is encountered during populating the diagram,
the error is ignored) because in some edge cases the errors are expected
to occur.

Release note (sql change): Statement diagnostics bundles now contain
output of `EXPLAIN (VEC)` and `EXPLAIN (VEC, VERBOSE)` commands for the
statements.
