+++ pkg/sql/BUILD.bazel
@@ -283,8 +283,6 @@ go_library(
         "//pkg/sql/catalog/tabledesc",
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/colexec",
-        "//pkg/sql/colexecerror",
-        "//pkg/sql/colexecop",
         "//pkg/sql/colflow",
         "//pkg/sql/contention",
         "//pkg/sql/covering",
@@ -378,7 +376,6 @@ go_library(
         "//pkg/util/timeutil",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
-        "//pkg/util/treeprinter",
         "//pkg/util/uint128",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_apd_v2//:apd",
