+++ pkg/sql/explain_bundle.go
@@ -162,9 +162,8 @@ func buildStatementBundle(
 	b.addStatement()
 	b.addOptPlans()
 	b.addExecPlan(planString)
-	// TODO(yuzefovich): consider adding some variant of EXPLAIN (VEC) output
-	// of the query to the bundle.
 	b.addDistSQLDiagrams()
+	b.addExplainVec()
 	traceJSON := b.addTrace()
 	b.addEnv(ctx)
 
@@ -309,6 +308,23 @@ func (b *stmtBundleBuilder) addDistSQLDiagrams() {
 			filename = fmt.Sprintf("distsql-%d-%s.html", i+1, d.typ)
 		}
 		b.z.AddFile(filename, contents)
+	}
+}
+
+func (b *stmtBundleBuilder) addExplainVec() {
+	for i, d := range b.plan.distSQLFlowInfos {
+		if len(d.explainVec) > 0 || len(d.explainVecVerbose) > 0 {
+			extra := ""
+			if len(b.plan.distSQLFlowInfos) > 1 {
+				extra = fmt.Sprintf("-%d-%s", i+1, d.typ)
+			}
+			if len(d.explainVec) > 0 {
+				b.z.AddFile(fmt.Sprintf("vec%s.txt", extra), strings.Join(d.explainVec, "\n"))
+			}
+			if len(d.explainVecVerbose) > 0 {
+				b.z.AddFile(fmt.Sprintf("vec%s-v.txt", extra), strings.Join(d.explainVecVerbose, "\n"))
+			}
+		}
 	}
 }
