+++ pkg/sql/explain_vec.go
@@ -12,21 +12,15 @@ package sql
 
 import (
 	"context"
-	"reflect"
-	"sort"
 
-	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/base"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descs"
-	"github.com/cockroachdb/cockroach/pkg/sql/colexecerror"
-	"github.com/cockroachdb/cockroach/pkg/sql/colexecop"
 	"github.com/cockroachdb/cockroach/pkg/sql/colflow"
 	"github.com/cockroachdb/cockroach/pkg/sql/execinfra"
-	"github.com/cockroachdb/cockroach/pkg/sql/execinfrapb"
 	"github.com/cockroachdb/cockroach/pkg/sql/physicalplan"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondatapb"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
-	"github.com/cockroachdb/cockroach/pkg/util/treeprinter"
 	"github.com/cockroachdb/errors"
 )
 
@@ -43,11 +37,6 @@ type explainVecNode struct {
 		// The current row returned by the node.
 		values tree.Datums
 	}
-}
-
-type flowWithNode struct {
-	nodeID roachpb.NodeID
-	flow   *execinfrapb.FlowSpec
 }
 
 func (n *explainVecNode) startExec(params runParams) error {
@@ -74,8 +63,7 @@ func (n *explainVecNode) startExec(params runParams) error {
 
 	distSQLPlanner.FinalizePlan(planCtx, physPlan)
 	flows := physPlan.GenerateFlowSpecs()
-	flowCtx := newFlowCtxForExplainPurposes(planCtx, params)
-	flowCtx.Cfg.ClusterID = &distSQLPlanner.rpcCtx.ClusterID
+	flowCtx := newFlowCtxForExplainPurposes(planCtx, params.p, &distSQLPlanner.rpcCtx.ClusterID)
 
 	// We want to get the vectorized plan which would be executed with the
 	// current 'vectorize' option. If 'vectorize' is set to 'off', then the
@@ -86,48 +74,29 @@ func (n *explainVecNode) startExec(params runParams) error {
 	if flowCtx.EvalCtx.SessionData.VectorizeMode == sessiondatapb.VectorizeOff {
 		return errors.New("vectorize is set to 'off'")
 	}
-
-	sortedFlows := make([]flowWithNode, 0, len(flows))
-	for nodeID, flow := range flows {
-		sortedFlows = append(sortedFlows, flowWithNode{nodeID: nodeID, flow: flow})
-	}
-	// Sort backward, since the first thing you add to a treeprinter will come last.
-	sort.Slice(sortedFlows, func(i, j int) bool { return sortedFlows[i].nodeID < sortedFlows[j].nodeID })
-	tp := treeprinter.NewWithStyle(treeprinter.CompactStyle)
-	root := tp.Child("│")
 	verbose := n.options.Flags[tree.ExplainFlagVerbose]
-	for _, flow := range sortedFlows {
-		node := root.Childf("Node %d", flow.nodeID)
-		opChains, cleanup, err := colflow.ConvertToVecTree(params.ctx, flowCtx, flow.flow, physPlan.LocalProcessors, !willDistribute)
-		defer cleanup()
-		if err != nil {
-			return err
-		}
-		// It is possible that when iterating over execinfra.OpNodes we will hit
-		// a panic (an input that doesn't implement OpNode interface), so we're
-		// catching such errors.
-		if err := colexecerror.CatchVectorizedRuntimeError(func() {
-			for _, op := range opChains {
-				formatOpChain(op, node, verbose)
-			}
-		}); err != nil {
-			return err
-		}
+	n.run.lines, err = colflow.ExplainVec(
+		params.ctx, flowCtx, flows, physPlan.LocalProcessors, verbose, willDistribute,
+	)
+	if err != nil {
+		return err
 	}
-	n.run.lines = tp.FormattedRows()
 	return nil
 }
 
-func newFlowCtxForExplainPurposes(planCtx *PlanningCtx, params runParams) *execinfra.FlowCtx {
+func newFlowCtxForExplainPurposes(
+	planCtx *PlanningCtx, p *planner, clusterID *base.ClusterIDContainer,
+) *execinfra.FlowCtx {
 	return &execinfra.FlowCtx{
 		NodeID:  planCtx.EvalContext().NodeID,
 		EvalCtx: planCtx.EvalContext(),
 		Cfg: &execinfra.ServerConfig{
-			Settings:       params.p.execCfg.Settings,
-			VecFDSemaphore: params.p.execCfg.DistSQLSrv.VecFDSemaphore,
+			Settings:       p.execCfg.Settings,
+			ClusterID:      clusterID,
+			VecFDSemaphore: p.execCfg.DistSQLSrv.VecFDSemaphore,
 		},
 		TypeResolverFactory: &descs.DistSQLTypeResolverFactory{
-			Descriptors: params.p.Descriptors(),
+			Descriptors: p.Descriptors(),
 		},
 		DiskMonitor: &mon.BytesMonitor{},
 	}
@@ -149,46 +118,6 @@ func newPlanningCtxForExplainPurposes(
 		p.result = tree.DNull
 	}
 	return planCtx
-}
-
-func shouldOutput(operator execinfra.OpNode, verbose bool) bool {
-	_, nonExplainable := operator.(colexecop.NonExplainable)
-	return !nonExplainable || verbose
-}
-
-func formatOpChain(operator execinfra.OpNode, node treeprinter.Node, verbose bool) {
-	seenOps := make(map[reflect.Value]struct{})
-	if shouldOutput(operator, verbose) {
-		doFormatOpChain(operator, node.Child(reflect.TypeOf(operator).String()), verbose, seenOps)
-	} else {
-		doFormatOpChain(operator, node, verbose, seenOps)
-	}
-}
-func doFormatOpChain(
-	operator execinfra.OpNode,
-	node treeprinter.Node,
-	verbose bool,
-	seenOps map[reflect.Value]struct{},
-) {
-	for i := 0; i < operator.ChildCount(verbose); i++ {
-		child := operator.Child(i, verbose)
-		childOpValue := reflect.ValueOf(child)
-		childOpName := reflect.TypeOf(child).String()
-		if _, seenOp := seenOps[childOpValue]; seenOp {
-			// We have already seen this operator, so in order to not repeat the full
-			// chain again, we will simply print out this operator's name and will
-			// not recurse into its children. Note that we print out the name
-			// unequivocally.
-			node.Child(childOpName)
-			continue
-		}
-		seenOps[childOpValue] = struct{}{}
-		if shouldOutput(child, verbose) {
-			doFormatOpChain(child, node.Child(childOpName), verbose, seenOps)
-		} else {
-			doFormatOpChain(child, node, verbose, seenOps)
-		}
-	}
 }
 
 func (n *explainVecNode) Next(runParams) (bool, error) {
