+++ pkg/sql/plan.go
@@ -267,6 +267,10 @@ var _ planNodeSpooled = &spoolNode{}
 type flowInfo struct {
 	typ     planComponentType
 	diagram execinfrapb.FlowDiagram
+	// explainVec and explainVecVerbose are only populated when collecting a
+	// statement bundle when the plan was vectorized.
+	explainVec        []string
+	explainVecVerbose []string
 	// flowsMetadata stores metadata from flows that will be used by
 	// execstats.TraceAnalyzer.
 	flowsMetadata *execstats.FlowsMetadata
